requirements.txt
pandas
transformers
scikit-learn
tqdm
accelerate
sentencepiece
protobuf
hf_xet
kaggle
bitsandbytes 
peft 
python-dotenv